Bumped version to 3.0.0-alpha
