+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "svelte-gantt",
-	"version": "2.0.0-alpha",
+	"version": "3.0.0-alpha",
 	"scripts": {
 		"build": "rollup -c",
 		"autobuild": "rollup -c -w",
