+++ public/gantt-default.css
@@ -94,7 +94,7 @@
     border-top: 1px solid #ddd;
 } */
 
-.column-header {
+.column-header-cell {
     border-right: #efefef 1px solid;
     border-bottom: #efefef 1px solid;
 }
