+++ src/column/index.ts
@@ -1,11 +1,11 @@
 import Column from './Column.svelte';
+import ColumnHeaderRow from './ColumnHeaderRow.svelte';
 import ColumnHeader from './ColumnHeader.svelte';
-import ColumnHeaders from './ColumnHeaders.svelte';
 import Columns from './Columns.svelte';
 
 export {
     Column,
+    ColumnHeaderRow,
     ColumnHeader,
-    ColumnHeaders,
     Columns
 };
